+++ src/firebase.ts
@@ -1,0 +1,36 @@
+import Constants from "expo-constants";
+import { getApps, initializeApp } from "firebase/app";
+import { getAuth, GoogleAuthProvider } from "firebase/auth";
+import { getFirestore } from "firebase/firestore";
+import { Platform } from "react-native";
+
+const firebaseConfig = {
+  apiKey: Constants.expoConfig?.extra?.firebaseApiKey ?? "",
+  authDomain: Constants.expoConfig?.extra?.firebaseAuthDomain ?? "",
+  projectId: Constants.expoConfig?.extra?.firebaseProjectId ?? "",
+  storageBucket: Constants.expoConfig?.extra?.firebaseStorageBucket ?? "",
+  messagingSenderId:
+    Constants.expoConfig?.extra?.firebaseMessagingSenderId ?? "",
+  appId: Constants.expoConfig?.extra?.firebaseAppId ?? "",
+};
+
+const app = getApps().length ? getApps()[0] : initializeApp(firebaseConfig);
+export const db = getFirestore(app);
+
+export const auth = (() => {
+  if (Platform.OS === "web") {
+    return getAuth(app);
+  } else {
+    const {
+      initializeAuth,
+      getReactNativePersistence,
+    } = require("firebase/auth");
+    const ReactNativeAsyncStorage =
+      require("@react-native-async-storage/async-storage").default;
+    return initializeAuth(app, {
+      persistence: getReactNativePersistence(ReactNativeAsyncStorage),
+    });
+  }
+})();
+
+export const googleProvider = new GoogleAuthProvider();