+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.2] - 2026-05-17
+
+### Added
+
+- Collection insertion on Firebase database.
+- Listing collections owned by user.
+
 ## [0.0.1] - 2026-05-16
 
 ### Added